[wip] Move some things out of big file.
