+++ src/LaunchDarkly/FeatureFlag.php
@@ -167,231 +167,6 @@ class FeatureFlag {
     }
 }
 
-class VariationOrRollout {
-    private static $LONG_SCALE = 0xFFFFFFFFFFFFFFF;
-
-    /** @var int */
-    private $_variation = null;
-    /** @var Rollout */
-    private $_rollout = null;
-
-    protected function __construct($variation, $rollout) {
-        $this->_variation = $variation;
-        $this->_rollout = $rollout;
-    }
-
-    public static function getDecoder() {
-        return function ($v) {
-            return new VariationOrRollout(
-                isset($v['variation']) ? $v['variation'] : null,
-                isset($v['rollout']) ? call_user_func(Rollout::getDecoder(), $v['rollout']) : null);
-        };
-    }
-
-    /**
-     * @return int
-     */
-    public function getVariation() {
-        return $this->_variation;
-    }
-
-    /**
-     * @return Rollout
-     */
-    public function getRollout() {
-        return $this->_rollout;
-    }
-
-    /**
-     * @param $user LDUser
-     * @param $_key string
-     * @param $_salt string
-     * @return int|null
-     */
-    public function variationIndexForUser($user, $_key, $_salt) {
-        if ($this->_variation !== null) {
-            error_log("Returning: $this->_variation");
-            return $this->_variation;
-        } else if ($this->_rollout !== null) {
-            $bucketBy = $this->_rollout->getBucketBy() === null ? "key" : $this->_rollout->getBucketBy();
-            $bucket = $this->bucketUser($user, $_key, $bucketBy, $_salt);
-            $sum = 0.0;
-            foreach ($this->_rollout->getVariations() as $wv) {
-                $sum += $wv->getWeight() / 100000.0;
-                if ($bucket < $sum) {
-                    return $wv->getVariation();
-                }
-            }
-        }
-        error_log("both fallthrough and variation are null!!");
-        return null;
-    }
-
-    /**
-     * @param $user LDUser
-     * @param $_key string
-     * @param $attr string
-     * @param $_salt string
-     * @return float
-     */
-    private function bucketUser($user, $_key, $attr, $_salt) {
-        $userValue = $user->getValueForEvaluation($attr);
-        $idHash = null;
-        if ($userValue != null) {
-            if (is_string($userValue)) {
-                $idHash = $userValue;
-                if ($user->getSecondary() !== null) {
-                    $idHash = $idHash . "." . $user->getSecondary();
-                }
-                $hash = substr(sha1($_key . "." . $_salt . "." . $idHash), 0, 15);
-                $longVal = base_convert($hash, 16, 10);
-                $result = $longVal / self::$LONG_SCALE;
-
-                return $result;
-            }
-        }
-        return 0.0;
-    }
-}
-
-class Clause {
-    private $_attribute = null;
-    private $_op = null;
-    private $_values = array();
-    private $_negate = false;
-
-    private function __construct($attribute, $op, array $values, $negate) {
-        $this->_attribute = $attribute;
-        $this->_op = $op;
-        $this->_values = $values;
-        $this->_negate = $negate;
-    }
-
-    public static function getDecoder() {
-        return function ($v) {
-            return new Clause($v['attribute'], $v['op'], $v['values'], $v['negate']);
-        };
-    }
-
-    /**
-     * @param $user LDUser
-     * @return bool
-     */
-    public function matchesUser($user) {
-        $userValue = $user->getValueForEvaluation($this->_attribute);
-//        error_log("user value: $userValue");
-        if ($userValue === null) {
-            error_log("null user value");
-            return false;
-        }
-        if (is_array($userValue)) {
-            error_log("uservalue is array");
-            foreach ($userValue as $element) {
-                if ($this->matchAny($element)) {
-                    return $this->_maybeNegate(true);
-                }
-            }
-            return $this->_maybeNegate(false);
-        } else {
-            error_log("else...");
-            return $this->_maybeNegate($this->matchAny($userValue));
-        }
-    }
-
-    /**
-     * @return null
-     */
-    public function getAttribute() {
-        return $this->_attribute;
-    }
-
-    /**
-     * @return null
-     */
-    public function getOp() {
-        return $this->_op;
-    }
-
-    /**
-     * @return array
-     */
-    public function getValues() {
-        return $this->_values;
-    }
-
-    /**
-     * @return boolean
-     */
-    public function isNegate() {
-        return $this->_negate;
-    }
-
-    /**
-     * @param $userValue
-     * @return bool
-     */
-    private function matchAny($userValue) {
-        foreach ($this->_values as $v) {
-            $result = Operators::apply($this->_op, $userValue, $v);
-            error_log("clause.matchany operator result for v: $v $result");
-            if ($result) {
-                error_log("true for $userValue");
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private function _maybeNegate($b) {
-        if ($this->_negate) {
-            return !$b;
-        } else {
-            return $b;
-        }
-    }
-}
-
-class Rule extends VariationOrRollout {
-    /** @var Clause[] */
-    private $_clauses = array();
-
-    protected function __construct($variation, $rollout, array $clauses) {
-        parent::__construct($variation, $rollout);
-        $this->_clauses = $clauses;
-    }
-
-    public static function getDecoder() {
-        return function ($v) {
-            return new Rule(
-                isset($v['variation']) ? $v['variation'] : null,
-                isset($v['rollout']) ? call_user_func(Rollout::getDecoder(), $v['rollout']) : null,
-                array_map(Clause::getDecoder(), $v['clauses']));
-        };
-    }
-
-    /**
-     * @param $user LDUser
-     * @return bool
-     */
-    public function matchesUser($user) {
-        foreach ($this->_clauses as $clause) {
-            if (!$clause->matchesUser($user)) {
-                error_log("false from rule.matchesuser with attr: " . $clause->getAttribute());
-                return false;
-            }
-        }
-        error_log("true from rule.matchesuser");
-        return true;
-    }
-
-    /**
-     * @return Clause[]
-     */
-    public function getClauses() {
-        return $this->_clauses;
-    }
-}
-
 class WeightedVariation {
     /** @var int */
     private $_variation = null;
